Refactor grammar.rs into trait-based dialect module
Split the monolithic grammar.rs (2808 lines, struct with static methods)
into a trait-based module directory:

- `grammar/mod.rs`: `Grammar` trait with `dispatch_statement` and
  `statement_keywords` as required methods, ~55 default methods for
  shared ANSI logic
- `grammar/ansi.rs`: `AnsiGrammar` struct implementing the trait
- `grammar/tsql.rs`: `TsqlGrammar` struct with 14 TSQL-specific methods
  and ANSI fallback delegation

`Parser` now holds `Box<dyn Grammar>` for dynamic dispatch, and
`DialectKind::parser()` wires the appropriate Grammar impl.

Also fixes `is_clause_keyword` to use `eq_ignore_ascii_case` instead of
heap-allocating `to_ascii_uppercase()`.

All 67 workspace tests pass, clippy clean, fmt clean.
