+++ crates/rigsql-parser/src/lib.rs
@@ -3,5 +3,5 @@ mod grammar;
 mod parser;
 
 pub use context::ParseContext;
-pub use grammar::Grammar;
+pub use grammar::{AnsiGrammar, Grammar, TsqlGrammar};
 pub use parser::{ParseError, Parser};
